routing-agent/pyproject.toml
[project]
name = "routing-agent"
version = "0.1.0"
description = "A routing agent that directs requests to appropriate specialized agents"
authors = [
    {name = "Your Name jashanpratap123@gmail.com"}
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
absl-py = "^2.2.1"
diff-match-patch = "^20241021"
google-adk = "^1.0.0"
google-cloud-bigquery = "^3.30.0"
google-genai = "^1.5.0"
pdfplumber = "^0.11.5"
pydantic = "^2.10.6"
requests = "^2.32.3"
tabulate = "^0.9.0"
scikit-learn = "^1.6.1"
google-cloud-aiplatform = { extras = [
  "adk",
  "agent-engines",
], version = "^1.93.0" }


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
